feat: supply event service and repository

+++ src/main/java/com/orderflow/orderflow_api/repositories/SupplyEventRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SupplyEventRepository extends CrudRepository<SupplyEvent,Integer> {
+public interface SupplyEventRepository extends CrudRepository<SupplyEvent, Long> {
+
 }
